+++ demo/components/StatusBar.tsx
@@ -65,7 +65,7 @@ export const StatusBar = (storyProps: StoryProps) => {
   }, progress < 100 ? 500 : null);
 
   return (
-    <L.Div _demoStory>
+    <L.Div style={{ paddingBottom: '24px' }} _demoStory>
       <L.H4 _title>StatusBar & StatusBarCustom</L.H4>
       <L.StatusBar
         data={data}